Move config to the core level. Don't use config class in config manager.
